+++ RME.rb
@@ -6467,6 +6467,7 @@ class Game_Temp
     Game_Temp.current_troop = 0
   end
   attr_accessor :last_used_item
+  attr_accessor :last_used_skill
 end
 
 
@@ -6537,7 +6538,8 @@ class Game_Battler < Game_BattlerBase
   # * Memorize item ID
   #--------------------------------------------------------------------------
   def use_item(item)
-    $game_temp.last_used_item = item.id
+    $game_temp.last_used_item = item.id if item.is_a?(RPG::Item)
+    $game_temp.last_used_skill = item.id if item.is_a?(RPG::Skill)
     old_use_item(item)
   end
 end
@@ -13265,6 +13267,7 @@ module RMECommands
     # end
 
     def last_used_item(); $game_temp.last_used_item; end
+    def last_used_skill(); $game_temp.last_used_skill; end
 
     append_commands
   end