+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Release (2.0.0)
 
+### Additions
+- Add command `last_used_skill` (@xvw, requested by @CuddleFox)
+
+### Bug fixes
+- Fix `last_used_item` to introduce distinction between skills and items (@xvw)
 - Fix followers displacement (@xvw, reported by @Erwill)
 - Fix `create_text_window` size inference (@xvw, reported by @AurelienDosSantos)
 - Great improvement on global performances (@xvw with @BlancLapin and @AurelienDosSantos 's help)